requirements.txt
einops==0.8.1
faster_whisper==1.2.1
ffmpeg_python==0.2.0
Flask==3.1.2
funasr==1.2.7
gradio==6.1.0
huggingface_hub==0.34.4
librosa==0.11.0
modelscope==1.31.0
numpy==2.3.5
onnxruntime==1.23.2
onnxruntime_gpu==1.23.2
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
packaging==25.0
pyannote.audio==3.1.1
pyannote.pipeline==3.0.1
pydub==0.25.1
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
rotary_embedding_torch==0.6.5
soundfile==0.13.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
tqdm==4.67.1